literature/requirements.txt
requests>=2.31
pyyaml>=6.0
sqlalchemy>=2.0
biopython>=1.83
python-Levenshtein>=0.25
